+++ src/main/kotlin/org/jetbrains/changelog/exceptions/MissingVersionException.kt
@@ -1,3 +1,3 @@
 package org.jetbrains.changelog.exceptions
 
-class MissingVersionException(version: String) : Exception("Version has no changelog: $version")
+class MissingVersionException(version: String) : Exception("Version is missing: $version")